requirements.txt
torch
transformers
accelerate
flask
langchain
langchain_community
transformers
torch
sentence-transformers
faiss-cpu
tqdm
pytesseract
Pillow
docx2txt
PyPDF2
openpyxl
pandas
argparse